# polyglot/packages/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Полиглот — vocabulary trainer API with session-based spaced repetition
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /words:
    get:
      operationId: listWords
      tags: [words]
      summary: List all words
      parameters:
        - in: query
          name: search
          schema:
            type: string
          description: Search in Russian, Polish, German, English fields
        - in: query
          name: sortBy
          schema:
            type: string
            enum: [frequency, createdAt, russian]
          description: Sort field
        - in: query
          name: limit
          schema:
            type: integer
          description: Page size
        - in: query
          name: offset
          schema:
            type: integer
          description: Page offset
      responses:
        "200":
          description: List of words
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WordListResponse"
    post:
      operationId: createWord
      tags: [words]
      summary: Create a new word
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WordInput"
      responses:
        "201":
          description: Created word
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Word"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /words/{id}:
    get:
      operationId: getWord
      tags: [words]
      summary: Get a word by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Word
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Word"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateWord
      tags: [words]
      summary: Update a word
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WordUpdate"
      responses:
        "200":
          description: Updated word
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Word"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteWord
      tags: [words]
      summary: Delete a word
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /words/{id}/mark-familiar:
    post:
      operationId: markWordFamiliar
      tags: [words]
      summary: Mark word as already partially known — skip Прописи and enter SRS directly
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Word added to SRS queue
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Word"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /words/{id}/restore:
    post:
      operationId: restoreWord
      tags: [words]
      summary: Restore a soft-deleted word from trash
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Restored word
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Word"
        "404":
          description: Not found
  /words/{id}/history:
    get:
      operationId: getWordHistory
      tags: [words]
      summary: Get training/review event history for a word
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Word event history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WordHistoryResponse"
  /words/bulk-import:
    post:
      operationId: bulkImportWords
      tags: [words]
      summary: Import multiple words at once
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkImportInput"
      responses:
        "200":
          description: Import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkImportResult"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /words/{id}/trace:
    post:
      operationId: recordWordTrace
      tags: [words]
      summary: Increment trace count for a word
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated word
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Word"
        "404":
          description: Word not found
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    Word:
      type: object
      required: [id, russian, createdAt]
      properties:
        id:
          type: integer
        russian:
          type: string
        polish:
          type: ["string", "null"]
        german:
          type: ["string", "null"]
        english:
          type: ["string", "null"]
        mnemonic:
          type: ["string", "null"]
        frequencyRank:
          type: ["integer", "null"]
          description: Frequency rank in academic dictionary (lower = more frequent)
        correctCount:
          type: integer
          description: Total correct answers across all languages
        hintCount:
          type: integer
          description: Total hints requested
        nextReviewAt:
          type: ["string", "null"]
          format: date-time
          description: Marker - null means new word, non-null means in SRS queue
        nextReviewSession:
          type: ["integer", "null"]
          description: Global session number when this word is next due for review
        reviewInterval:
          type: integer
          description: Number of times reviewed correctly (index into intervals array)
        wordType:
          type: ["string", "null"]
          description: Usage context of the word (any custom tag value)
        wordGroup:
          type: ["string", "null"]
          description: Mnemonic group label — words sharing the same memory technique/image
        semanticGroup:
          type: ["string", "null"]
          description: Semantic group label — synonyms / words interchangeable in context
        priority:
          type: integer
          description: Priority for Прописи ordering (0 = normal, 1 = high — words that were answered wrong)
          default: 0
        consecutiveCorrect:
          type: integer
          description: Consecutive correct SRS answers — resets to 0 on wrong answer
          default: 0
        traceCount:
          type: integer
          description: How many times this word has been completed in Прописи (trace) sessions
          default: 0
        graduatedAt:
          type: ["string", "null"]
          format: date-time
          description: When set — word is fully learned and removed from active review rotation
        deletedAt:
          type: ["string", "null"]
          format: date-time
          description: Soft-delete timestamp; null means active, non-null means in trash
        createdAt:
          type: string
          format: date-time
    WordInput:
      type: object
      required: [russian]
      properties:
        russian:
          type: string
          minLength: 1
        polish:
          type: string
        german:
          type: string
        english:
          type: string
        mnemonic:
          type: string
        frequencyRank:
          type: integer
        wordType:
          type: string
          description: Word type value (any custom tag value)
        wordGroup:
          type: string
          description: Mnemonic group label
        semanticGroup:
          type: string
          description: Semantic group label (synonyms / interchangeable words)
    WordUpdate:
      type: object
      properties:
        russian:
          type: string
          minLength: 1
        polish:
          type: string
        german:
          type: string
        english:
          type: string
        mnemonic:
          type: string
        frequencyRank:
          type: integer
        wordType:
          type: string
          description: Word type value (any custom tag value)
        wordGroup:
          type: string
          description: Mnemonic group label
        semanticGroup:
          type: string
          description: Semantic group label (synonyms / interchangeable words)
    WordListResponse:
      type: object
      required: [words, total]
      properties:
        words:
          type: array
          items:
            $ref: "#/components/schemas/Word"
        total:
          type: integer
    BulkImportInput:
      type: object
      required: [words]
      properties:
        words:
          type: array
          items:
            $ref: "#/components/schemas/WordInput"
    BulkImportResult:
      type: object
      required: [imported, skipped, errors]
      properties:
        imported:
          type: integer
        skipped:
          type: integer
        errors:
          type: array
          items:
            type: string
...
    WordHistoryEvent:
      type: object
      required: [id, eventType, sessionNumber, createdAt]
      properties:
        id:
          type: integer
        eventType:
          type: string
          description: "correct | wrong | hint | review_correct | review_wrong"
        sessionNumber:
          type: integer
        createdAt:
          type: string
          format: date-time
    WordHistoryResponse:
      type: object
      required: [wordId, events]
      properties:
        wordId:
          type: integer
        events:
          type: array
          items:
            $ref: "#/components/schemas/WordHistoryEvent"
...