# polyglot/packages/api-zod/src/generated/types/bulkImportResult.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Полиглот — vocabulary trainer API with session-based spaced repetition
 * OpenAPI spec version: 0.1.0
 */

export interface BulkImportResult {
  imported: number;
  skipped: number;
  errors: string[];
}