# polyglot/packages/api-zod/src/generated/types/errorResponse.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Полиглот — vocabulary trainer API with session-based spaced repetition
 * OpenAPI spec version: 0.1.0
 */

export interface ErrorResponse {
  error: string;
}