# polyglot/packages/api-zod/src/generated/types/wordUpdate.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Полиглот — vocabulary trainer API with session-based spaced repetition
 * OpenAPI spec version: 0.1.0
 */

export interface WordUpdate {
  /** @minLength 1 */
  russian?: string;
  polish?: string;
  german?: string;
  english?: string;
  mnemonic?: string;
  frequencyRank?: number;
  /** Word type value (any custom tag value) */
  wordType?: string;
  /** Mnemonic group label */
  wordGroup?: string;
  /** Semantic group label (synonyms / interchangeable words) */
  semanticGroup?: string;
}